22-01-23_RNA-seq_First_Code
The first code for performing RNA-seq according to the Computational genomics book. 

For Github purposes, this isn't great for visually showing how the code works, or what graphs are produced. I will need to consider writing my code in R-markdown format, or possibly in some form of open access format here on GitHUB, so that all of the code chunks work on gitHUB, and produce the graphs. Although it might just be easier to write everything in an R-markdown format, and then just including images of the figures after the code chunks that produce figures. The same would apply to tables of data which are produced.