22-01-23_RNA-seq_First_Code
This is the first code chunk I have produced for RNA-seq. It pretty much is just pulled straight from the Computational Genomics online book, aside from a few functions and streamlining parts. 

It is clear that this format is not suitable for showing the code on GitHub. I will need to move to using Markdown formats for producing the code, but will still provide a raw 'R' form of the end code sections so that they can be copied and used for other works.